Rename class to App and move it to seperate folder
